VLBACal.py: Modified function VLBAUVFITS to copy cross- AND auto-correations.  By default only cross-correlations are copied.
Quack.c:
In Obit task Quack added check for NX table.  If not present, Quack
sets an error and returns nicely.

These changes were made to correct for bugs detected when trying to 
process a single-source data file with the VLBA pipeline.  Ultimately,
we discovered this is not possible because single-source files do not have
the source id random axis.  The pipeline expects this axis.