I discovered that r241 inadvertently removed good code that was added in r240. I have merged the latest revision, r243, with the code added in r240.  I have paid little attention to some recent changes I made pertaining to the 'check' mode.  These should be easy to fix / add later if needed.
In addition... 

Task input paging has been turned off when debug mode is active (this is
accomplished via the pydoc module).

The doBand argument in the call to VLBAImageTargets from VLBAContPipe.py has
been set to -1 to avoid a BP table error in the final imaging run.  (In this
final run, the averaged & calibrated data is used to generate images; since
this data has already had the bandpass calibration applied, it has no BP table.
Therefore, when doBand was true, a BP table error was being generated.)