* VLBACal.py *
Added a function to write an HTML report of the pipeline data reduction.  This
report includes project and single-source metadata and output files.

Added outfiles object to hold information on files written by pipeline.  The
outfiles object is a mix of dictionaries and lists that separate file
information depending on whether it is a multi-source/project file or a
single-source file.  outfiles is used when writing output files in the HTML
report.  Since files are only added to outfiles when they are created (if the
appropriate function, below, is called), generating complete reports when
_rerunning_ subsets of the pipeline requires that files generated on previous
runs be saved in a pickle 'jar'.    

Three functions have been written to manipulate outfiles.

VLBAAddOutFile(...) stores file information in outfiles.  
VLBASaveOutFiles(...) saves the outfiles object as a pickle file.  
VLBAFetchOutFiles(...) retrieves the outfiles pickle jar and restores outfiles.  
Checks to verify that files exist before adding them to outfiles.

Renamed VLBAReportTargets -> VLBASrcMetadata

* VLBAContPipe.py *

Changed input parameter doReport -> doMetadata

Added calls to outfiles functions where appropriate.