In function VLBAIDILoad, I have changed the code to create a symbolic link to the input uvfits file. FITLD then follows this link when loading data from disk. FITLD can only handle file names with no more than 48 characters.  Some archive files have more characters.  Sym links serve as a work-around for this problem.  The link is deleted after data is loaded into AIPS.  I call 'uuidgen' from the command line to obtain a UUID that is used for the link file name. This should prevent concurrent pipeline processes from trampling on each other when creating and destroying sym links.
Added functions to query, download, and handle metadata from the NRAO archive:

VLBAQueryArchive - Query the archive for a subset of data between an input
start and stop time.

VLBAParseArchiveResponse - Parse the archive query response and store the
metadata in a list of dictionaries.

VLBADownloadArchiveFile - Initiate a download of a specific file from the
archive.

VLBASummarizeArchiveResponse - Print a table summary of the archive query
response

VLBAGetParms - Create a list of pipeline input parameters using the metadata in
the archive query response

VLBAPrepare - Prepare pipeline for processing.  This is a wrapper function that
invokes all the above functions to setup the pipeline for processing and ends
by telling the user how to start the pipeline.