* VLBAContPipeWrap.py
I have added a new bookkeeping file that I call the "pipeline processing
record".  The file is stored in my user directory on file host because I want
it backed up daily.  The processing record is a Python pickle file that stores
a list of archive file dictionaries for files that have been pipeline
processed.  When the pipeline finishes properly, the pipeline wrapper will
append an archive file dictionary to the end of the processing record list.
Dictionaries in the processing record contain an additional key: 'pipe.STATUS'.
The value of this key is currently either 'check' or 'archive' to denote that
the corresponding output data set is either at the check (validation) stage or
the archive stage.

* PipeUtil.py

Modified function SummarizeArchiveResponse to compare each row of the summary
table with the contents of the pipeline processing record.  If the file is
already in the processing record, append an astrisk to the summary table row.