* PipeUtil.py
Added functions to get, put, and extract from the pipeline record list.  Added
logging basic config so logging messages are displayed when functions are run
manually.

* VLBAContPipe.py

Removed './' from the beginning of several file names, at the request of John
Benson.  Added VOTable to outfiles before writing the VOTable; this way the
VOTable refers to itself as one of the project data files; John requested this.
Added check for consistency between outfiles and the current working directory;
if not consistent, raise and exception; added the DataProductException for this
purpose.

* VLBACal.py

Added './' to the beginning of all lwpla.outfile values. Removed some debugging
messages that were not supposed to be committed. Added 'fileSetID' to metadata
object and to VOTable at John B's request. Modified VLBAFetchOutFiles to
automatically remove from outfiles any files that do not exist in the current
directory. Added function VLBAValidOutfiles to compare outfiles with the
current directory and report on differences.

* VLBAContPipeWrap.py

Added CalledProcessError exception handling.  This will handle exceptions from
the subprocess.check_call that invokes the pipeline function.  Hopefully, if
the pipeline crashes, the wrapper will continue with the next project.