* PipeUtil.py
Added validate(...) function to perform data validation steps after a dataset
has been verified.  Reformatted some docstrings.

* VLBAContPipe.py

Only generate contour plots for sources that are contained in either
parms["targets"], parms["contCals"], or both.  This prevents non-detected (and
non-imaged) sources from being plotted.

* VLBACal.py

Changed fileSetID value in project metadata dictionary to use on the last 2
digits of the year.  This is the format of the date returned by the NRAO
archive, and the format used when the data set directory is created.

* VLBAContPipeWrap.py

Return to starting directory (cwd) when an IOError or
subprocess.CalledProcessError is raised.  In other words, when the pipeline
crashes, return to the starting directory before kicking off another pipeline
process.