Added automated FITS IDI data loading to VLBA pipeline.
* PipeUtil.py

Had to change the 'format' parameter value in the archive query HTTP request.
This caused the archive to return some additional files that were not coming
through previously.  Added a filter to remove files with format=VLBAset.

* VLBAContPipeWrap.py

Changed some of the command line options. Added option for loading IDI files.